+++ src/mandrill/test.clj
@@ -4,9 +4,10 @@
 
 (defn env-token
   "Clojure.test fixture that sets the Mandrill API token for all tests
-  using the environment variable MANDRILL_DEV_TOKEN.
+  using the environment variable linked to the supplied keyword.
 
-  Use like: (clojure.test/use-fixtures :once t/env-token)"
-  [test-fn]
-  (m/with-token (:mandrill-dev-token e/env)
-    (test-fn)))
+  Use like: (clojure.test/use-fixtures :once (t/env-token :mandrill-dev-token))"
+  [k]
+  (fn [test-fn]
+    (m/with-token (k e/env)
+      (test-fn))))