+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1 (10/09/2014)
+
+- Replace `:mandrill-dev-token` with `:mandrill-token`
+- Use an env variable, `:mandrill-token`, as a default.
+
 ## 0.1.0 (10/09/2014)
 
 - Initial release.